ASoC: wcd9xxx: fix the NULL pointer dereference causing the crash
If the probe of the device is unsuccessful then the wcd9xxx
structure is unallocated but is not initialized to NULL.
If some operation like reset comes, it would dereference the wcd9xxx
structure leading to the target crash.
Fix by intializing the client data to NULL after it is unallocated.

Change-Id: Ida5b598887a12ad36fc759a67d4cc9c570f359ec
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>