ASoC: q6asm: Fix the memory leak during the SSR
If a session is released during the SSR, then the memory unmap
command to DSP fails. This results in not releasing up the memory
block shared among apps and the DSP.
Fix by checking if the session has been invalidated during the SSR
and free up the memory during the close command.

Change-Id: Iaa47aa5a3a8b59cabe37b17a056658ae54c8626f
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>