mm: Add notifier framework for showing memory
There are many drivers in the kernel which can hold on
to lots of memory. It can be useful to dump out all those
drivers at key points in the kernel. Introduct a notifier
framework for dumping this information. When the notifiers
are called, drivers can dump out the state of any memory
they may be using.

Change-Id: If577aaebcb4587ef8fdabdb6cf209a406a1afd4d
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>