msm-core: Ensure polling occurs on schedule
MSM core driver is supposed to update power cost and cpu temperature on a
fixed interval in addition to when there is a large change in temperature.
There is a poll_ms parameter that is set to 50ms by default that should
cause the driver to poll for updates every 50ms. It has been observed that
this polling can stop for up to several minutes.

The issue is identified with the optimization in the polling code which
stalls polling if temperature didn't change between two consecutive
readings and polling isn't started until after a large temperature change
triggers the polling thread. Fix issue by removing the optimization.

Change-Id: I12a5eb8470bd9655cb834a19a72bda3caf02a714
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>