ASoC: msm: qdsp6v2: fix null pointer access during SSR
While SSR in progress, CMD_CLOSE is sent to asm driver
and at the same time ac->apr handle is set to NULL
as part of RESET_EVENTS callback to this driver.
As both contexts are not synchronized, sometimes
ac->apr handle structure is dereferenced after it
is set to NULL. Fix the issue by having synchronization
between the two contexts.

Change-Id: Id3ee4b7290acef857005055a142bd474e3c038ef
Signed-off-by: Laxminath Kasam <lkasam@codeaurora.org>