CHROMIUM: net: wireless: bcmdhd: no retry on rxglom superframe error
Aborting the current read attempt on the superframe also removes the packet
from the pipeline. Retries should not be attempted on the next packet since
it would not be a superframe(either a superframe descriptor or other data
packet) and should not be handled by dhdsdio_rxglom.

BUG=chrome-os-partner:45478
TEST=manual test with iperf on ryu

Change-Id: I3f6d7ff8c6028ed5b6a978ff8d9c8fa099127070
Signed-off-by: Franky Lin <frankyl@broadcom.com>
Reviewed-on: https://chromium-review.googlesource.com/309909
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Dmitry Shmidt <dimitrysh@google.com>